Split single segment and multi segment functionality and related code refactoring
